Make the following changes:
,

Find this element:

  <Header>

with the text content:

    FBHomeServicesAboutContactToggle Menu

Relevant files:

    - https://6000-firebase-studio-1747324005069.cluster-oayqgyglpfgseqclbygurw4xd4.cloudworkstations.dev/_next/static/chunks/node_modules_next_dist_compiled_2ce9398a._.js

And change it as follows:

from far left, the name is not FB, its F.B/c and it should have the small glowing orange dot on the left side of it   we need to center all the navigaton button, and you should add back all the addtional pages you made in the start.   the theme button dont work.