look nothing looks centerd.
and maybe we should move the F.B/C logo with the dot out and place it in the left side of the header and not have it in the dock.

and another important factor how will this work on any devices ? will it adjust size? or will it stay the same.

and the icons should not go out of the dock this much.

i need your help im no expert, just use best practises to make this work.